streamline client version handling upgrade to 0.54.4 client
