+++ build.gradle
@@ -11,8 +11,8 @@ repositories {
 }
 
 dependencies {
-    compile 'io.crate:crate-client:0.54.3'
-    testCompile "io.crate:crate-testing:0.1.1"
+    compile 'io.crate:crate-client:0.54.4'
+    testCompile "io.crate:crate-testing:0.2.0"
     testCompile 'org.hamcrest:hamcrest-all:1.3'
     testCompile 'junit:junit:4.11'
     testCompile ('com.carrotsearch.randomizedtesting:randomizedtesting-runner:2.1.11') {
@@ -36,52 +36,6 @@ sourceSets {
             srcDir 'src/test/java'
         }
     }
-}
-
-ext {
-    downloadDir = new File(rootDir, 'downloads')
-    crate_version = '0.54.3'
-}
-
-
-def download (String url, String name){
-    downloadDir.mkdirs()
-    def dest = new File(downloadDir, name)
-    ant.get(
-            src: url,
-            verbose: true,
-            usetimestamp: true,
-            skipexisting: true,
-            dest: dest
-    )
-    return dest;
-}
-
-def extractTopfolder(File src, String trg){
-    // delete target folder. Otherwise old files will stick around and lead to strange errors
-    new File(trg).delete()
-
-    copy {
-        includeEmptyDirs = false
-        FileTree ft
-        if(src.name.endsWith('.zip')) {
-            ft = zipTree(src)
-        }
-        if(src.name.endsWith('.tar.gz')) {
-            ft = tarTree(src)
-        }
-        from(ft) {
-            eachFile { details ->
-                details.path =
-                        details.path.substring(details.relativePath.segments[0].length())
-            }
-        }
-        into trg
-    }
-}
-
-task deleteCrateDownloadFolder(type: Delete) {
-    delete "${project.rootDir}/parts/crate"
 }
 
 task getVersion ( dependsOn: 'classes') << {
@@ -295,10 +249,6 @@ test {
     // show standard out and standard error of the test JVM(s) on the console
     testLogging.showStandardStreams = true
     outputs.upToDateWhen { false }
-}
-
-clean {
-    dependsOn deleteCrateDownloadFolder
 }
 
 idea {
